refactor: shift plot into main relic tab
